# Moving stack trace capturing away from application thread