--- a/src/Elastic.Apm/Model/Span.cs
+++ b/src/Elastic.Apm/Model/Span.cs
@@ -20,6 +20,7 @@ namespace Elastic.Apm.Model
 {
 	internal class Span : ISpan
 	{
+		private readonly IApmServerInfo _apmServerInfo;
 		private readonly Lazy<SpanContext> _context = new Lazy<SpanContext>();
 		private readonly ICurrentExecutionSegmentsContainer _currentExecutionSegmentsContainer;
 		private readonly Transaction _enclosingTransaction;
@@ -28,22 +29,6 @@ namespace Elastic.Apm.Model
 		private readonly IApmLogger _logger;
 		private readonly Span _parentSpan;
 		private readonly IPayloadSender _payloadSender;
-
-		/// <summary>
-		/// In some cases capturing the stacktrace in <see cref="End" /> results in a stack trace which is not very useful.
-		/// In such cases we capture the stacktrace on span start.
-		/// These are typically async calls - e.g. capturing stacktrace for outgoing HTTP requests in the
-		/// System.Net.Http.HttpRequestOut.Stop
-		/// diagnostic source event produces a stack trace that does not contain the caller method in user code - therefore we
-		/// capture the stacktrace is .Start
-		/// </summary>
-		private readonly StackFrame[] _stackFrames;
-
-		/// <summary>
-		/// Captures the sample rate of the agent when this span was created.
-		/// </summary>
-		[JsonProperty("sample_rate")]
-		internal double SampleRate { get; }
 
 		// This constructor is meant for deserialization
 		[JsonConstructor]
@@ -100,20 +85,14 @@ namespace Elastic.Apm.Model
 				{
 					enclosingTransaction.SpanCount.IncrementStarted();
 
-					if (captureStackTraceOnStart)
-					{
-						var stackTrace = new StackTrace(true);
-						try
-						{
-							// I saw EnhancedStackTrace throwing exceptions in some environments
-							// therefore we try-catch and fall back to a non-demystified call stack.
-							_stackFrames = new EnhancedStackTrace(stackTrace).GetFrames();
-						}
-						catch
-						{
-							_stackFrames = stackTrace.GetFrames();
-						}
-					}
+					// In some cases capturing the stacktrace in End() results in a stack trace which is not very useful.
+					// In such cases we capture the stacktrace on span start.
+					// These are typically async calls - e.g. capturing stacktrace for outgoing HTTP requests in the
+					// System.Net.Http.HttpRequestOut.Stop
+					// diagnostic source event produces a stack trace that does not contain the caller method in user code - therefore we
+					// capture the stacktrace is .Start
+					if (captureStackTraceOnStart && ConfigSnapshot.StackTraceLimit != 0 && ConfigSnapshot.SpanFramesMinDurationInMilliseconds != 0)
+						RawStackTrace = new StackTrace(true);
 				}
 			}
 			else
@@ -127,13 +106,12 @@ namespace Elastic.Apm.Model
 		}
 
 		private bool _isEnded;
-		private readonly IApmServerInfo _apmServerInfo;
 
 		[MaxLength]
 		public string Action { get; set; }
 
 		[JsonIgnore]
-		private IConfigSnapshot ConfigSnapshot => _enclosingTransaction.ConfigSnapshot;
+		internal IConfigSnapshot ConfigSnapshot => _enclosingTransaction.ConfigSnapshot;
 
 		/// <summary>
 		/// Any other arbitrary data captured by the agent, optionally provided by the user.
@@ -185,6 +163,20 @@ namespace Elastic.Apm.Model
 		[MaxLength]
 		[JsonProperty("parent_id")]
 		public string ParentId { get; set; }
+
+		/// <summary>
+		/// This holds the raw stack trace that was captured when the span either started or ended (depending on the parameter
+		/// passed to the .ctor)
+		/// This will be turned into an elastic stack trace and sent to APM Server in the <see cref="StackTrace" /> property
+		/// </summary>
+		[JsonIgnore]
+		internal StackTrace RawStackTrace;
+
+		/// <summary>
+		/// Captures the sample rate of the agent when this span was created.
+		/// </summary>
+		[JsonProperty("sample_rate")]
+		internal double SampleRate { get; }
 
 		[JsonIgnore]
 		internal bool ShouldBeSentToApmServer => IsSampled && !_isDropped;
@@ -298,38 +290,40 @@ namespace Elastic.Apm.Model
 
 				// Spans are sent only for sampled transactions so it's only worth capturing stack trace for sampled spans
 				// ReSharper disable once CompareOfFloatsByEqualityOperator
-				if (ConfigSnapshot.StackTraceLimit != 0 && ConfigSnapshot.SpanFramesMinDurationInMilliseconds != 0)
-				{
-					if (Duration >= ConfigSnapshot.SpanFramesMinDurationInMilliseconds
-						|| ConfigSnapshot.SpanFramesMinDurationInMilliseconds < 0)
-					{
-						if (_stackFrames == null)
-						{
-							StackFrame[] trace;
-							var stackTrace = new StackTrace(true);
-							try
-							{
-								// I saw EnhancedStackTrace throwing exceptions in some environments
-								// therefore we try-catch and fall back to a non-demystified call stack.
-								trace = new EnhancedStackTrace(stackTrace).GetFrames();
-							}
-							catch
-							{
-								trace = stackTrace.GetFrames();
-							}
+				if (ConfigSnapshot.StackTraceLimit != 0 && ConfigSnapshot.SpanFramesMinDurationInMilliseconds != 0 && RawStackTrace == null
+					&& (Duration >= ConfigSnapshot.SpanFramesMinDurationInMilliseconds
+						|| ConfigSnapshot.SpanFramesMinDurationInMilliseconds < 0))
+					RawStackTrace = new StackTrace(true);
 
-							StackTrace = StacktraceHelper.GenerateApmStackTrace(trace,
-								_logger,
-								ConfigSnapshot, _apmServerInfo, $"Span `{Name}'");
-						}
-						else
-						{
-							StackTrace = StacktraceHelper.GenerateApmStackTrace(_stackFrames,
-								_logger,
-								ConfigSnapshot, _apmServerInfo, $"Span `{Name}'");
-						}
-					}
-				}
+				// if (Duration >= ConfigSnapshot.SpanFramesMinDurationInMilliseconds
+				// 	|| ConfigSnapshot.SpanFramesMinDurationInMilliseconds < 0)
+				// {
+				// 	if (StackFrames == null)
+				// 	{
+				// 		StackFrame[] trace;
+				// 		var stackTrace = new StackTrace(true);
+				// 		try
+				// 		{
+				// 			// I saw EnhancedStackTrace throwing exceptions in some environments
+				// 			// therefore we try-catch and fall back to a non-demystified call stack.
+				// 			trace = new EnhancedStackTrace(stackTrace).GetFrames();
+				// 		}
+				// 		catch
+				// 		{
+				// 			trace = stackTrace.GetFrames();
+				// 		}
+				//
+				// 		StackTrace = StacktraceHelper.GenerateApmStackTrace(trace,
+				// 			_logger,
+				// 			ConfigSnapshot, _apmServerInfo, $"Span `{Name}'");
+				// 	}
+				// 	else
+				// 	{
+				// 		StackTrace = StacktraceHelper.GenerateApmStackTrace(StackFrames,
+				// 			_logger,
+				// 			ConfigSnapshot, _apmServerInfo, $"Span `{Name}'");
+				// 	}
+				// }
 
 				_payloadSender.QueueSpan(this);
 			}