--- a/test/Elastic.Apm.Tests/Mocks/MockPayloadSender.cs
+++ b/test/Elastic.Apm.Tests/Mocks/MockPayloadSender.cs
@@ -8,19 +8,27 @@ using System.Linq;
 using System.Threading.Tasks;
 using System.Timers;
 using Elastic.Apm.Api;
+using Elastic.Apm.Filters;
+using Elastic.Apm.Logging;
 using Elastic.Apm.Metrics;
 using Elastic.Apm.Model;
 using Elastic.Apm.Report;
+using Elastic.Apm.ServerInfo;
 
 namespace Elastic.Apm.Tests.Mocks
 {
 	internal class MockPayloadSender : IPayloadSender
 	{
+		internal readonly List<Func<ISpan, ISpan>> SpanFilters = new List<Func<ISpan, ISpan>>();
 		private readonly List<IError> _errors = new List<IError>();
 		private readonly object _lock = new object();
 		private readonly List<IMetricSet> _metrics = new List<IMetricSet>();
 		private readonly List<ISpan> _spans = new List<ISpan>();
 		private readonly List<ITransaction> _transactions = new List<ITransaction>();
+
+		public MockPayloadSender(IApmLogger logger = null)
+			=> SpanFilters.Add(
+				new SpanStackTraceCapturingFilter(logger ?? new NoopLogger(), new MockApmServerInfo(new ElasticVersion(7, 10, 0, null))).Filter);
 
 		private TaskCompletionSource<ITransaction> _transactionTaskCompletionSource = new TaskCompletionSource<ITransaction>();
 
@@ -28,7 +36,7 @@ namespace Elastic.Apm.Tests.Mocks
 
 		public Error FirstError => _errors.First() as Error;
 
-		internal void ResetTransactionTaskCompletionSource() => _transactionTaskCompletionSource = new TaskCompletionSource<ITransaction>();
+		public MetricSet FirstMetric => _metrics.First() as MetricSet;
 
 		/// <summary>
 		/// The 1. Span on the 1. Transaction
@@ -36,11 +44,9 @@ namespace Elastic.Apm.Tests.Mocks
 		public Span FirstSpan => _spans.First() as Span;
 
 		public Transaction FirstTransaction =>
-			 Transactions.First() as Transaction;
+			Transactions.First() as Transaction;
 
 		public IReadOnlyList<IMetricSet> Metrics => CreateImmutableSnapshot(_metrics);
-
-		public MetricSet FirstMetric => _metrics.First() as MetricSet;
 
 		public IReadOnlyList<ISpan> Spans => CreateImmutableSnapshot(_spans);
 
@@ -51,11 +57,7 @@ namespace Elastic.Apm.Tests.Mocks
 		{
 			get
 			{
-
-				var timer = new Timer
-				{
-					Interval = 1000
-				};
+				var timer = new Timer { Interval = 1000 };
 
 				timer.Enabled = true;
 				timer.Start();
@@ -82,6 +84,8 @@ namespace Elastic.Apm.Tests.Mocks
 			}
 		}
 
+		internal void ResetTransactionTaskCompletionSource() => _transactionTaskCompletionSource = new TaskCompletionSource<ITransaction>();
+
 		public void QueueError(IError error) => _errors.Add(error);
 
 		public virtual void QueueTransaction(ITransaction transaction)
@@ -92,8 +96,11 @@ namespace Elastic.Apm.Tests.Mocks
 
 		public void QueueSpan(ISpan span)
 		{
-			lock(_lock)
+			lock (_lock)
+			{
+				foreach (var filter in SpanFilters) span = filter(span);
 				_spans.Add(span);
+			}
 		}
 
 		public void QueueMetrics(IMetricSet metricSet) => _metrics.Add(metricSet);