Addresses https://github.com/elastic/apm-agent-dotnet/issues/1039

`new StackTrace(true)` still happens on the app thread, but the demystification and the conversion to our internal stack trace model happens right before serialization on the payloadsender thread.

Benchmarks (I worked with [this benchmark](https://github.com/elastic/apm-agent-dotnet/blob/2182be81af478390111061e3edee280c4cafa3b9/test/Elastic.Apm.PerfTests/AspNetCorePerf/AspNetCoreLoadTestWithAgent.cs))

Current master:

```
|                             Method |     Mean |    Error |   StdDev |   Median |
|----------------------------------- |---------:|---------:|---------:|---------:|
| WebRequestWithDbCallsAndCustomSpan | 31.96 ms | 3.569 ms | 10.52 ms | 24.93 ms |
```

With this PR:

```
|                             Method |     Mean |     Error |    StdDev |   Median |
|----------------------------------- |---------:|----------:|----------:|---------:|
| WebRequestWithDbCallsAndCustomSpan | 8.325 ms | 0.2154 ms | 0.6214 ms | 8.126 ms |
```

So from the throughput perspective of the monitored app this seems like an improvement.

Potential issues: 
- With this we moved more work to the `PayloadSenderV2` (or to be precise to the thread that handles every span after `.End()` is called on it)
- If people do `apm.Tracer.CurrentSpan` then the `StackTrace` property will be `null` on it, which wasn't the case previously (although to be fair the story there wasn't very nice, because in some cases it was filled when the span was ended (that's the default) and in other cases (e.g. HTTP and DB spans) it was filled when the span started). On the positive side, if people use filters, then at that point the property will be already filled.