@gregkalapos 2020-11-25
I did further testing:  first with `1.7.0`, then with branch of the [PR](https://github.com/elastic/apm-agent-dotnet/pull/1052). Here is the transaction duration:

![image](https://user-images.githubusercontent.com/1091853/100272340-3acd6200-2f5b-11eb-947a-2ecdd5d998a1.png)

I tested [with this sample app from the repo](https://github.com/elastic/apm-agent-dotnet/blob/master/sample/SampleAspNetCoreApp/Controllers/HomeController.cs#L47) - only change I did is that I removed the outgoing HTTP request, since that's not deterministic. This test creates 4 EF Core DB spans in each request.

Test scenario: generating full load with Netling for 10 minutes.

Avg Transaction duration:
- `1.7.0`: well above 100ms, often above 150ms
- With the [PR](https://github.com/elastic/apm-agent-dotnet/pull/1052): mostly between 40ms and 50ms.